Fixed the insertion order of the polls, and can thus use the index_serial for id instead of the old method. This will allow for easier future development. However, care need to be take that the index_serial be reindex if a poll is removed in order to not create a gap in ids.
Made !find give 9 results instead of 5 due to discords new embeds. These embeds can now have 3 fields per line. And seems to always inline unless `inline=False` is used.
So by going for 9 results we get a 3x3 grid of results which looks ok? might be changing this in the future.
Side note* The embed colors are messed up in dark mode currently, making the `name` fields grey on black. Might have experiment with embeds to solve this if discord does not.

Added a message when !getpoll is done fetching all the polls